# Optimize `TargetNode`'s set operations