Part of: https://github.com/tuist/tuist/issues/1093

### Short description 📝

While profiling the workspace generator (#1093) the `TargetNode` equality check was identified as an area that could be simplified to gain some performance improvements.

- `TargetNode` is one of the graph node types used in Tuist
- It is a node that points to a target (normal target within a project)
- All nodes form a `Graph` in Tuist and there are several operations performed on the graph that require traversing the nodes and their dependencies
- These operations involve storing nodes in a `Set`
- This means indirectly each node's equality and hashing methods are called for each node during traversals
- The `TargetNode` was checking its associated `target` for equality each time

### Solution 📦

- In the current context of how nodes are used, checking `target` equality is unnecessary as nodes need only be identified by type, path and name.

- The equality check can be simplified and updated to match the hashing method which only accounted for `name` and `path`.

### Implementation 👩‍💻👨‍💻

- [x] Update `TargetNode` equality check
- [x] Update change log

### Test Plan 🛠

- Verify all automated tests pass

### Profiling Results ⏱

From profiling #1093, a number of components measurements improve as many of them directly or indirectly perform graph traversal operations which this change improves. 

| Components              | Duration     | Percentage | Delta    | 
| ----------------------- | ------------ | ---------- | -------- |
| WorkspaceGenerator      | 7.99 s       | 100.00%    | -81.45%  |
| ProjectGenerator        | 7.97 s       | 99.75%     | -81.49%  |
| ProjectFileElements     | 3.84 s       | 48.06%     | -81.94%  |
| TargetGenerator         | 3.68 s       | 46.06%     | -82.71%  |
| BuildPhaseGenerator     | 3.44 s       | 43.05%     | 3.93%    |
| LinkGenerator           | 133.24 ms    | 1.67%      | -99.25%  |
| ConfigGenerator         | 96.80 ms     | 1.21%      | 4.73%    |